requirements.txt
Django==2.1.2
django-treebeard==4.3
djangorestframework==3.9.0
Pillow==5.3.0
